+++ lib/legion/extensions/builders/actors.rb
@@ -25,7 +25,7 @@ module Legion
               Legion::Logging.warn "[Actors] constant #{actor_class} not defined, skipping" if defined?(Legion::Logging)
               next
             end
-            Legion::Logging.info "[Actors] built actor: #{actor_class}" if defined?(Legion::Logging)
+            log.info "[Actors] built actor: #{actor_class}" if defined?(Legion::Logging)
             @actors[actor_name.to_sym] = {
               extension:      lex_class.to_s.downcase,
               extension_name: extension_name,